% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caseI.R
\name{caseI}
\alias{caseI}
\title{This function runs the simulation and plots the figure for case I.}
\usage{
caseI(beta)
}
\arguments{
\item{beta}{The value of beta, which is between 0 and 1. In the paper, beta is chosen as 0.3 and 0.4 respectively.}
}
\description{
This function runs the simulation and plots the figure for case I.
}
