% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLAT.R
\name{CLAT}
\alias{CLAT}
\title{The main routine for CLAT}
\usage{
CLAT(pv, q)
}
\arguments{
\item{pv}{A vector of pvalues.}

\item{q}{The designated mFDR level}
}
\value{
R: The total number of rejection.

SigInd: A vector of 0-1, indicating whether the corresponding hypothesis is rejected (1) or is not rejected (0).


}
\description{
This function loads a vector of p-values and q, the designated mFDR level.
}
\examples{



p <- 1000

X <- c( rnorm(100, -2, 1), rnorm(100, 2, 1), rnorm(800, 0, ) )

q <- 0.05

## Right-sided test
p.right <- pnorm( X, lower.tail=FALSE )
clat.right <- CLAT( p.right, q)
rejind <- which( clat.right$SigInd!=0 )

## Right-sided test
p.left <- pnorm( X, lower.tail=TRUE )
clat.left <- CLAT( p.left, q)
rejind <- which( clat.left$SigInd!=0 )


## Two-sided test
p.right <- pnorm( X, lower.tail=FALSE )
clat.right <- CLAT( p.right, q)

p.left <- pnorm( X, lower.tail=TRUE )
clat.left <- CLAT( p.left, q)
rejind <- c( which( clat.left$SigInd!=0 ), which(clat.right$SigInd!=0) ) 
}
