% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLAT_example.R
\name{CLAT_example}
\alias{CLAT_example}
\title{This function runs the simulation for the settings in the paper.}
\usage{
CLAT_example(case, beta, numSim = 1000)
}
\arguments{
\item{case}{Case number, the possible choices are "I", "II","III", "IV".}

\item{beta}{The value of beta, which is between 0 and 1. In the paper, beta is chosen as 0.3 and 0.4 respectively.}
}
\description{
This function runs the simulation for the settings in the paper.
}
