% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caseI.R
\name{caseI}
\alias{caseI}
\title{This function runs the simulation and plots the figure for case I.}
\usage{
caseI(beta, numSim = 1000)
}
\arguments{
\item{beta}{The value of beta, which is between 0 and 1. In the paper, beta is chosen as 0.3 and 0.4 respectively.}

\item{numSim}{The number of replications with a default choice of 1000.}
}
\description{
This function runs the simulation and plots the figure for case I.
}
