% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caseIV.R
\name{caseIV}
\alias{caseIV}
\title{This function runs the simulation and plots the figure for case IV.}
\usage{
caseIV(beta, numSim = 1000)
}
\arguments{
\item{beta}{The value of beta, which is between 0 and 1. In the paper, beta is chosen as 0.3 and 0.4 respectively.}

\item{numSim}{The number of replications with a default choice of 1000.}
}
\description{
This function runs the simulation and plots the figure for case IV.
}
